requirements.txt
openai
typer
rich
python-dotenv
prompt_toolkit
